Refactor: update directory structure to support additional root layout
