+++ src/app/(routes)/page.module.scss
@@ -1,4 +1,4 @@
-@import "./styles/colors.scss";
+@import "../_styles/colors.scss";
 
 .headingExample {
   color: $primary-color;
