+++ src/app/(routes)/page.test.tsx
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 import { render, screen } from "@testing-library/react";
-import Home from "@/app/page";
+import Home from "@/app/(routes)/page";
 
 it("Home", () => {
   render(<Home />);
