+++ src/app/_styles/colors.scss
